refactor(#792): split block_tests into base + model-ref/disk-package suite (sweep)
block_tests.rs 957 -> 477, + block_tests_more.rs 488 (both under cap; off debt).
The model-ref + disk-package end-to-end tests move to the sibling; delay_block
becomes pub(super), imported via super::tests. project tests (17) pass, zero warnings.
